refactor: move exception handlers
