+++ packages/api-server/src/sessions/sessions.controller.ts
@@ -17,7 +17,6 @@ import { UpdateSessionDto } from './dto/update-session.dto';
 import { UserId } from '../users/decorators/user-id.decorator';
 import { JwtGuard } from '../auth/guards/jwt.guard';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
-import { FilesController } from '../files/files.controller';
 
 @ApiTags('Session')
 @ApiBearerAuth()
@@ -27,7 +26,6 @@ export class SessionsController {
     private readonly sessionsService: SessionsService,
     private readonly filesService: FilesService,
     private readonly sessionFilesService: SessionFilesService,
-    private readonly filesController: FilesController,
   ) {}
 
   @Post('create')
@@ -95,7 +93,7 @@ export class SessionsController {
 
   async validateFileIds(userId: number, fileIds: number[]) {
     for (const fileId of fileIds) {
-      await this.filesController.getFileAsUser(fileId, userId);
+      await this.filesService.getFileAsUser(fileId, userId);
     }
   }
 }
